// src/shared/api/models/getApiCustomers200CustomersItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiCustomers200CustomersItemEmail } from './getApiCustomers200CustomersItemEmail';
import type { GetApiCustomers200CustomersItemCountryCode } from './getApiCustomers200CustomersItemCountryCode';
import type { GetApiCustomers200CustomersItemCountryISOCode } from './getApiCustomers200CustomersItemCountryISOCode';
import type { GetApiCustomers200CustomersItemGiftCardsItem } from './getApiCustomers200CustomersItemGiftCardsItem';
import type { GetApiCustomers200CustomersItemBundleOffersItem } from './getApiCustomers200CustomersItemBundleOffersItem';

export type GetApiCustomers200CustomersItem = {
    id: number;
    name: string;
    phoneNumber: string;
    email?: GetApiCustomers200CustomersItemEmail;
    outstandingAmount: number;
    countryCode?: GetApiCustomers200CustomersItemCountryCode;
    countryISOCode?: GetApiCustomers200CustomersItemCountryISOCode;
    bonus?: number;
    giftCards: GetApiCustomers200CustomersItemGiftCardsItem[];
    bundleOffers?: GetApiCustomers200CustomersItemBundleOffersItem[];
};